runable clean code
